% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main_functions.R
\name{drawAnnotations}
\alias{drawAnnotations}
\title{Extract annotation informations from the annotation package org.Hs.eg.db}
\usage{
drawAnnotations(label = "Annotations")
}
\arguments{
\item{label}{character label of the track, default "Annotations"}

\item{file_path}{character absolute path to the file containing snp
informations}
}
\value{
a ggplot track
}
\description{
Extract annotation informations from the annotation package org.Hs.eg.db
}
\examples{
current_chr <- "chr12"
current_range <- setStudyRange(27950000, 28735000)
annot_track <- drawAnnotations("My Genes")
}

