% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main_functions.R
\name{drawArchs}
\alias{drawArchs}
\title{Prepare tracks to plot them}
\usage{
drawArchs(ranges_list, highlight_range_list = NULL)
}
\arguments{
\item{ranges_list}{list of GenomicRanges generated with the functions
\code{getDataOverview}}

\item{highlight_range_list}{list of GenomicRange generated with the function
\code{setHighLight}}
}
\value{
list of 3 lists containing GenomicRanges representing data in the
current study range : arch, circos, lncrna
}
\description{
Prepare tracks to plot them
}
\examples{
current_chr <- "chr12"
my.data <- loadChrData()
current_range <- setStudyRange(27950000, 28735000)
my.ranges <- getDataOverview()
specific_range1 <- setHighLight(28111017,28127138,"alpha")
specific_range2 <- setHighLight(28284682,28287682,"alpha")
specific_range3 <- setHighLight(28284682,28287682,"color")

arch <- drawArchs(my.ranges$arch, list(specific_range1,specific_range2,
specific_range3))
tracks(arch) + xlim(current_range)
}

