% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main_functions.R
\name{drawSNPs}
\alias{drawSNPs}
\title{Extract SNP informations from a flat file and build the resulting track}
\usage{
drawSNPs(file_path, label = "SNPs")
}
\arguments{
\item{file_path}{character absolute path to the file containing snp
informations}

\item{label}{character label of the track, default "SNPs"}
}
\value{
a ggplot track
}
\description{
Extract SNP informations from a flat file and build the resulting track
}
\examples{
snps_list = "~/Workspace/COLLAB/PTHLH_snps"
current_chr <- "chr12"
current_range <- setStudyRange(27950000, 28735000)
snps_track <- drawSNPs(snps_list, "My SNPs")
}

