% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main_functions.R
\name{getDataOverview}
\alias{getDataOverview}
\title{Check data availability and extract relevant data under the form of list of
GenomicRanges
After setting the current chromosome and study range}
\usage{
getDataOverview()
}
\value{
list of 3 lists containing GenomicRanges representing data in the
current study range : arch, circos, lncrna
}
\description{
Check data availability and extract relevant data under the form of list of
GenomicRanges
After setting the current chromosome and study range
}
\examples{
current_chr <- "chr12"
my.data <- loadChrData()
current_range <- setStudyRange(27950000, 28735000)
my.ranges <- getDataOverview()
}

