% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main_functions.R
\name{loadChrData}
\alias{loadChrData}
\title{Download data for a current_chromosome
After setting current_chr variable : current_chr <- "chr1"}
\usage{
loadChrData()
}
\value{
list elements containing all data available for a particular
chromosome
}
\description{
Download data for a current_chromosome
After setting current_chr variable : current_chr <- "chr1"
}
\examples{
current_chr <- "chr12"
my.data <- loadChrData()
}

