% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main_functions.R
\name{mergeRanges}
\alias{mergeRanges}
\title{Merge range in a single one}
\usage{
mergeRanges(ranges, label)
}
\arguments{
\item{ranges}{vector of GenomicRanges generated with the functions
\code{getDataOverview} you want to merge}

\item{label}{character label of the merged range}
}
\value{
a GenomicRange
}
\description{
Merge range in a single one
}
\examples{
current_chr <- "chr12"
my.data <- loadChrData()
current_range <- setStudyRange(27950000, 28735000)
my.ranges <- getDataOverview()
k562_chiapet <- mergeRanges(c(my.ranges$arch$sgp_k562_lane13,
my.ranges$arch$sgp_k562_lane24), label = "K562 ChiA-PET SGP")
}

