% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main_functions.R
\name{setHighLight}
\alias{setHighLight}
\title{Enable to set the zone to higlight}
\usage{
setHighLight(current_start, current_stop, method)
}
\arguments{
\item{current_start}{integer start of the range}

\item{current_stop}{integer end of the range}

\item{method}{character defining the way to highlight a specific zone (alpha
or color)}
}
\value{
a GenomicRange
}
\description{
Enable to set the zone to higlight
}
\examples{
specific_range <- setHighLight(28111017,28127138,"alpha")
}

