% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/main_functions.R
\name{setStudyRange}
\alias{setStudyRange}
\title{Enable to set the current study range}
\usage{
setStudyRange(current_start, current_stop)
}
\arguments{
\item{current_start}{integer start of the range}

\item{current_stop}{integer end of the range}
}
\value{
a GenomicRange
}
\description{
Enable to set the current study range
}
\examples{
current_range <- setStudyRange(27950000, 28735000)
}

